video-2-runtime-inline-tools/02-agent-with-tools/requirements.txt
bedrock-agentcore>=0.1.0
boto3>=1.34.0
strands-agents>=0.1.0
